+++ src/Compilers/Core/Portable/Diagnostic/WellKnownDiagnosticTags.cs
@@ -1,8 +1,6 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using Loretta.CodeAnalysis.Diagnostics;
-
 namespace Loretta.CodeAnalysis
 {
     /// <summary>
@@ -16,11 +14,6 @@ namespace Loretta.CodeAnalysis
         public const string Unnecessary = nameof(Unnecessary);
 
         /// <summary>
-        /// Indicates that the diagnostic is related to edit and continue.
-        /// </summary>
-        public const string EditAndContinue = nameof(EditAndContinue);
-
-        /// <summary>
         /// Indicates that the diagnostic is related to build.
         /// </summary>
         public const string Build = nameof(Build);
@@ -31,30 +24,8 @@ namespace Loretta.CodeAnalysis
         public const string Compiler = nameof(Compiler);
 
         /// <summary>
-        /// Indicates that the diagnostic can be used for telemetry
-        /// </summary>
-        public const string Telemetry = nameof(Telemetry);
-
-        /// <summary>
         /// Indicates that the diagnostic is not configurable, i.e. it cannot be suppressed or filtered or have its severity changed.
         /// </summary>
         public const string NotConfigurable = nameof(NotConfigurable);
-
-        /// <summary>
-        /// Indicates that the diagnostic is related to an exception thrown by a <see cref="DiagnosticAnalyzer"/>.
-        /// </summary>
-        public const string AnalyzerException = nameof(AnalyzerException);
-
-        /// <summary>
-        /// Indicates that the diagnostic is an obsolete diagnostic with a custom ID
-        /// specified by the 'DiagnosticId' property on 'ObsoleteAttribute'.
-        /// </summary>
-        public const string CustomObsolete = nameof(CustomObsolete);
-
-        /// <summary>
-        /// Indicates that the diagnostic is a compilation end diagnostic reported
-        /// from a compilation end action.
-        /// </summary>
-        public const string CompilationEnd = nameof(CompilationEnd);
     }
 }