+++ src/Compilers/Core/Portable/Syntax/GreenNodeExtensions.cs
@@ -108,18 +108,7 @@ namespace Loretta.CodeAnalysis
 
         public static TNode WithDiagnosticsGreen<TNode>(this TNode node, DiagnosticInfo[]? diagnostics) where TNode : GreenNode
         {
-            return (TNode)node.SetDiagnostics(diagnostics);
-        }
-
-        public static TNode WithoutDiagnosticsGreen<TNode>(this TNode node) where TNode : GreenNode
-        {
-            var current = node.GetDiagnostics();
-            if (current == null || current.Length == 0)
-            {
-                return node;
-            }
-
-            return (TNode)node.SetDiagnostics(null);
+            return (TNode) node.SetDiagnostics(diagnostics);
         }
     }
 }