+++ src/Compilers/Core/Portable/Syntax/SyntaxTree.cs
@@ -1,9 +1,7 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using System;
 using System.Collections.Generic;
-using System.Collections.Immutable;
 using System.Diagnostics.CodeAnalysis;
 using System.Text;
 using System.Threading;
@@ -19,12 +17,6 @@ namespace Loretta.CodeAnalysis
     public abstract class SyntaxTree
     {
         /// <summary>
-        /// Cached value for empty <see cref="DiagnosticOptions"/>.
-        /// </summary>
-        protected internal static readonly ImmutableDictionary<string, ReportDiagnostic> EmptyDiagnosticOptions =
-            ImmutableDictionary.Create<string, ReportDiagnostic>(CaseInsensitiveComparison.Comparer);
-
-        /// <summary>
         /// The path of the source document file.
         /// </summary>
         /// <remarks>
@@ -32,17 +24,6 @@ namespace Loretta.CodeAnalysis
         /// The path shall not be null.
         /// 
         /// The file doesn't need to exist on disk. The path is opaque to the compiler.
-        /// The only requirement on the path format is that the implementations of 
-        /// <see cref="SourceReferenceResolver"/>, <see cref="XmlReferenceResolver"/> and <see cref="MetadataReferenceResolver"/> 
-        /// passed to the compilation that contains the tree understand it.
-        /// 
-        /// Clients must also not assume that the values of this property are unique
-        /// within a Compilation.
-        /// 
-        /// The path is used as follows:
-        ///    - When debug information is emitted, this path is embedded in the debug information.
-        ///    - When resolving and normalizing relative paths in #r, #load, #line/#ExternalSource, 
-        ///      #pragma checksum, #ExternalChecksum directives, XML doc comment include elements, etc.
         /// </remarks>
         public abstract string FilePath { get; }
 
@@ -175,86 +156,6 @@ namespace Loretta.CodeAnalysis
         /// The values are not affected by line mapping directives (<c>#line</c>).
         /// </returns>
         public abstract FileLinePositionSpan GetLineSpan(TextSpan span, CancellationToken cancellationToken = default);
-
-        /// <summary>
-        /// Gets the location in terms of path, line and column after applying source line mapping directives 
-        /// (<c>#line</c> in C# or <c>#ExternalSource</c> in VB). 
-        /// </summary>
-        /// <param name="span">Span within the tree.</param>
-        /// <param name="cancellationToken">Cancellation token.</param>
-        /// <returns>
-        /// A valid <see cref="FileLinePositionSpan"/> that contains path, line and column information.
-        /// 
-        /// If the location path is mapped the resulting path is the path specified in the corresponding <c>#line</c>,
-        /// otherwise it's <see cref="SyntaxTree.FilePath"/>.
-        /// 
-        /// A location path is considered mapped if the first <c>#line</c> directive that precedes it and that 
-        /// either specifies an explicit file path or is <c>#line default</c> exists and specifies an explicit path.
-        /// </returns>
-        public abstract FileLinePositionSpan GetMappedLineSpan(TextSpan span, CancellationToken cancellationToken = default);
-
-        /// <summary>
-        /// Returns the visibility for the line at the given position.
-        /// </summary>
-        /// <param name="position">The position to check.</param>
-        /// <param name="cancellationToken">The cancellation token.</param> 
-        public virtual LineVisibility GetLineVisibility(int position, CancellationToken cancellationToken = default) => LineVisibility.Visible;
-
-        /// <summary>
-        /// Gets a FileLinePositionSpan for a TextSpan and the information whether this span is considered to be hidden or not. 
-        /// FileLinePositionSpans are used primarily for diagnostics and source locations.
-        /// This method combines a call to GetLineSpan and IsHiddenPosition.
-        /// </summary>
-        /// <param name="span"></param>
-        /// <param name="isHiddenPosition">Returns a boolean indicating whether this span is considered hidden or not.</param>
-        /// <remarks>This function is being called only in the context of sequence point creation and therefore interprets the 
-        /// LineVisibility accordingly (BeforeFirstRemappingDirective -> Visible).</remarks>
-        internal virtual FileLinePositionSpan GetMappedLineSpanAndVisibility(TextSpan span, out bool isHiddenPosition)
-        {
-            isHiddenPosition = GetLineVisibility(span.Start) == LineVisibility.Hidden;
-            return GetMappedLineSpan(span);
-        }
-
-        /// <summary>
-        /// Returns a path for particular location in source that is presented to the user. 
-        /// </summary>
-        /// <remarks>
-        /// Used for implementation of <see cref="System.Runtime.CompilerServices.CallerFilePathAttribute"/> 
-        /// or for embedding source paths in error messages.
-        /// 
-        /// Unlike Dev12 we do account for #line and #ExternalSource directives when determining value for 
-        /// <see cref="System.Runtime.CompilerServices.CallerFilePathAttribute"/>.
-        /// </remarks>
-        internal string GetDisplayPath(TextSpan span, SourceReferenceResolver? resolver)
-        {
-            var mappedSpan = GetMappedLineSpan(span);
-            if (resolver == null || mappedSpan.Path.IsEmpty())
-            {
-                return mappedSpan.Path;
-            }
-
-            return resolver.NormalizePath(mappedSpan.Path, baseFilePath: mappedSpan.HasMappedPath ? FilePath : null) ?? mappedSpan.Path;
-        }
-
-        /// <summary>
-        /// Returns a line number for particular location in source that is presented to the user. 
-        /// </summary>
-        /// <remarks>
-        /// Used for implementation of <see cref="System.Runtime.CompilerServices.CallerLineNumberAttribute"/> 
-        /// or for embedding source line numbers in error messages.
-        /// 
-        /// Unlike Dev12 we do account for #line and #ExternalSource directives when determining value for 
-        /// <see cref="System.Runtime.CompilerServices.CallerLineNumberAttribute"/>.
-        /// </remarks>
-        internal int GetDisplayLineNumber(TextSpan span)
-             // display line numbers are 1-based
-             => GetMappedLineSpan(span).StartLinePosition.Line + 1;
-
-        /// <summary>
-        /// Are there any hidden regions in the tree?
-        /// </summary>
-        /// <returns>True if there is at least one hidden region.</returns>
-        public abstract bool HasHiddenRegions();
 
         /// <summary>
         /// Returns a list of the changed regions between this tree and the specified tree. The list is conservative for