+++ src/Compilers/Core/Portable/Syntax/SyntaxTreeExtensions.cs
@@ -99,19 +99,5 @@ namespace Loretta.CodeAnalysis
             }
             return (n1 == n2) ? -1 : n + 1;
         }
-
-        /// <summary>
-        /// Returns <c>true</c> if the provided position is in a hidden region inaccessible to the user.
-        /// </summary>
-        public static bool IsHiddenPosition(this SyntaxTree tree, int position, CancellationToken cancellationToken = default)
-        {
-            if (!tree.HasHiddenRegions())
-            {
-                return false;
-            }
-
-            var lineVisibility = tree.GetLineVisibility(position, cancellationToken);
-            return lineVisibility == LineVisibility.Hidden || lineVisibility == LineVisibility.BeforeFirstLineDirective;
-        }
     }
 }