+++ src/Compilers/Core/Test/Utilities/Diagnostics/DiagnosticDescription.cs
@@ -124,7 +124,7 @@ namespace Loretta.CodeAnalysis.Test.Utilities
             _effectiveSeverityOpt = includeEffectiveSeverity ? d.Severity : (DiagnosticSeverity?) null;
 
             DiagnosticWithInfo dinfo = null;
-            if (d.Code == 0 || d.Descriptor.CustomTags.Contains(WellKnownDiagnosticTags.CustomObsolete))
+            if (d.Code == 0)
             {
                 _code = d.Id;
                 _errorCodeType = typeof(string);