+++ src/Compilers/Core/Test/Utilities/TestBase.cs
@@ -19,38 +19,10 @@ namespace Loretta.Test.Utilities
     /// <summary>
     /// Base class for all unit test classes.
     /// </summary>
-    public abstract class TestBase : IDisposable
+    public abstract class TestBase
     {
-        private TempRoot _temp;
-
         protected TestBase()
         {
-        }
-
-        public static string GetUniqueName()
-        {
-            return Guid.NewGuid().ToString("D");
-        }
-
-        public TempRoot Temp
-        {
-            get
-            {
-                if (_temp == null)
-                {
-                    _temp = new TempRoot();
-                }
-
-                return _temp;
-            }
-        }
-
-        public virtual void Dispose()
-        {
-            if (_temp != null)
-            {
-                _temp.Dispose();
-            }
         }
 
         #region Diagnostics