+++ src/Compilers/Lua/Portable/Syntax/LuaSyntaxTree.cs
@@ -303,35 +303,6 @@ namespace Loretta.CodeAnalysis.Lua
         }
 
         /// <summary>
-        /// Gets the location in terms of path, line and column after applying source line mapping directives (<c>#line</c>).
-        /// </summary>
-        /// <param name="span">Span within the tree.</param>
-        /// <param name="cancellationToken">Cancellation token.</param>
-        /// <returns>
-        /// <para>A valid <see cref="FileLinePositionSpan"/> that contains path, line and column information.</para>
-        /// <para>
-        /// If the location path is mapped the resulting path is the path specified in the corresponding <c>#line</c>,
-        /// otherwise it's <see cref="SyntaxTree.FilePath"/>.
-        /// </para>
-        /// <para>
-        /// A location path is considered mapped if the first <c>#line</c> directive that precedes it and that
-        /// either specifies an explicit file path or is <c>#line default</c> exists and specifies an explicit path.
-        /// </para>
-        /// </returns>
-        public override FileLinePositionSpan GetMappedLineSpan(TextSpan span, CancellationToken cancellationToken = default) =>
-            GetLineSpan(span, cancellationToken);
-
-        /// <inheritdoc/>
-        public override LineVisibility GetLineVisibility(int position, CancellationToken cancellationToken = default) =>
-            LineVisibility.Visible;
-
-        /// <summary>
-        /// Gets a boolean value indicating whether there are any hidden regions in the tree.
-        /// </summary>
-        /// <returns>True if there is at least one hidden region.</returns>
-        public override bool HasHiddenRegions() => false;
-
-        /// <summary>
         /// Gets a <see cref="Location"/> for the specified text <paramref name="span"/>.
         /// </summary>
         public override Location GetLocation(TextSpan span) => new SourceLocation(this, span);